Fix to #9455 - Query/Logging: log contents of compiled code (e.g. materializers inside a shaper) as part of the query plan
Shapers (especially projection shapers) contain useful information that gets compiled out of the query plan. For debugging purposes we stash the uncompiled expression on the Shaper as well so we can print the complete query plan.
We only do that if the query plans are getting logged to avoid unnecessary bloating of Shapers when that extra information doesn't get used.